benchmarks/requirements.txt
httpx>=0.27.0
e2b>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tabulate>=0.9.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
